Arrumando Header

+++ ft_strlcat.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcat.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: coder <coder@student.42.fr>                +#+  +:+       +#+        */
+/*   By: gmoreira <gmoreira@student.42sp.org.br     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/04/11 01:54:35 by gmoreira          #+#    #+#             */
-/*   Updated: 2022/05/13 21:34:43 by coder            ###   ########.fr       */
+/*   Created: 2022/05/14 20:14:13 by gmoreira          #+#    #+#             */
+/*   Updated: 2022/05/14 20:14:15 by gmoreira         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlcat(char *dst, const char *src, size_t size)
+size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
 	size_t	index_src;
 	size_t	index_dest;
