+++ libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gmoreira <gmoreira@student.42sp.org.br>    +#+  +:+       +#+        */
+/*   By: coder <coder@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/06 22:37:46 by gmoreira          #+#    #+#             */
-/*   Updated: 2022/05/12 23:24:57 by gmoreira         ###   ########.fr       */
+/*   Updated: 2022/05/14 19:50:36 by coder            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ void	*ft_memchr(const void *str, int c, size_t n);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
 void	*ft_memmove(void *dest, const void *src, size_t n);
 int		ft_strlcpy(char *dest, const char *src, size_t size);
-int		ft_strlcat(char *dest, const char *src, size_t size);
+size_t	ft_strlcat(char *dest, const char *src, size_t size);
 char	*ft_substr(const char *s, unsigned int start, size_t len);
 int		ft_memcmp(const void *str1, const void *str2, size_t num );
 int		ft_strncmp(const char *str1, const char *str2,
